added mmap count charts
